+++ app/concepts/stellar_base/deposit_requests/deposit_request_policy.rb
@@ -1,0 +1,14 @@
+module StellarBase
+  module DepositRequests
+    class DepositRequestPolicy
+
+      def initialize(_, _)
+      end
+
+      def create?
+        StellarBase.included_module?(:deposit)
+      end
+
+    end
+  end
+end