+++ app/representers/application_representer.rb
@@ -1,3 +1,0 @@
-class ApplicationRepresenter < Representable::Decorator
-  include Representable::JSON
-end