+++ app/representers/stellar_base/application_representer.rb
@@ -1,0 +1,5 @@
+module StellarBase
+  class ApplicationRepresenter < Representable::Decorator
+    include Representable::JSON
+  end
+end