+++ app/representers/stellar_base/deposit_request_representer.rb
@@ -1,0 +1,13 @@
+module StellarBase
+  class DepositRequestRepresenter < ApplicationRepresenter
+
+    property :deposit_address, as: :how
+    property :eta
+    property :min_amount
+    property :max_amount
+    property :fee_fixed
+    property :fee_percent
+    property :extra_info
+
+  end
+end