+++ app/services/stellar_base/gen_memo_for.rb
@@ -1,0 +1,17 @@
+module StellarBase
+  class GenMemoFor
+
+    LENGTH = 8
+
+    def self.call(klass)
+      loop do
+        memo = GenRandomString.(length: LENGTH)
+        if !klass.exists?(memo: memo)
+          return memo
+          break
+        end
+      end
+    end
+
+  end
+end