+++ app/twins/stellar_base/deposit_request_twin.rb
@@ -1,0 +1,13 @@
+module StellarBase
+  class DepositRequestTwin < ApplicationTwin
+
+    property :deposit_address
+    property :eta
+    property :min_amount
+    property :max_amount
+    property :fee_fixed
+    property :fee_percent
+    property :extra_info
+
+  end
+end