+++ config/routes.rb
@@ -1,5 +1,6 @@
 StellarBase::Engine.routes.draw do
   resources :bridge_callbacks, only: [:create], defaults: { format: "json" }
   get "/withdraw" => "withdraw#create", as: :withdraw
+  get "/deposit" => "deposit#create", as: :deposit
 end